remove bonus part

+++ ft_printf/ft_printf.c
@@ -6,7 +6,7 @@
 /*   By: ptippaya <ptippaya@student.42bangkok.co    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/06/15 13:03:21 by ptippaya          #+#    #+#             */
-/*   Updated: 2022/06/16 19:04:31 by ptippaya         ###   ########.fr       */
+/*   Updated: 2022/09/03 17:06:18 by ptippaya         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -23,11 +23,11 @@ int	ft_printf_format(char c, va_list arg)
 	else if (c == 'd' || c == 'i')
 		return (ft_putdec(va_arg(arg, int)));
 	else if (c == 'u')
-		return (ft_putudec(va_arg(arg, size_t)));
+		return (ft_putudec(va_arg(arg, unsigned int)));
 	else if (c == 'x')
-		return (ft_puthexl(va_arg(arg, size_t)));
+		return (ft_puthexl(va_arg(arg, unsigned int)));
 	else if (c == 'X')
-		return (ft_puthexu(va_arg(arg, size_t)));
+		return (ft_puthexu(va_arg(arg, unsigned int)));
 	else if (c == '%')
 		return (ft_putchar('%'));
 	return (0);
